combined wkb functions to one file
